improve checking for files to delete

+++ Optimage/Optimage/Controllers/ImageUploadController.cs
@@ -10,36 +10,56 @@ namespace Optimage.Controllers
 {
     public class ImageUploadController : Controller
     {
+        // Setting a variable for control the content root folder of solution
         private readonly IWebHostEnvironment Environment;
-        
         public ImageUploadController(IWebHostEnvironment _environment)
         {
             Environment = _environment;
         }
+
+        /* --- Startup action --- */
+        /**
+            Contains the procedures need to be activated when the project is launched:
+            - Auto clearing image directory from files and folders
+            - View result is a list of uploaded files (running the project, by default, the result is empty)
+        **/
         public IActionResult ImageUpload()
         {
             DirectoryInfo folder = new DirectoryInfo(Path.Combine(this.Environment.ContentRootPath, "Images"));
             foreach (FileInfo file in folder.EnumerateFiles())
             {
-                // exist !
-                file.Delete();
+                if (file.Exists)
+                {
+                    file.Delete();
+                }
             }
-
+            foreach (DirectoryInfo directory in folder.EnumerateDirectories())
+            {
+                if (directory.Exists)
+                {
+                    directory.Delete();
+                }
+            }
             var items = GetFiles();
             return View(items);
         }
 
+ 
+        /* --- Uploading action --- */
+        /**
+            Saves image upload directory. If missing, create a folder.
+            With using file stream creates file in choosen directory.
+            Uploading process it's just copying form file in folder.
+            A message of uploading result is displayed below the form and list of files becomes updated.
+        **/
         [HttpPost]
         public IActionResult ImageUpload(IFormFile uploadedImage)
         {
-            /// Save image upload directory. If missing, create a folder
             string appPath = Path.Combine(this.Environment.ContentRootPath, "Images");
             if (!Directory.Exists(appPath))
             {
                 Directory.CreateDirectory(appPath);
             }
-
-            /// Creates a file object and saves it to the FileStream. Then view output uploaded file
             try
             {
                 string fileName = Path.GetFileName(uploadedImage.FileName);
@@ -53,11 +73,15 @@ namespace Optimage.Controllers
             {
                 ViewBag.Message = "ERROR: " + ex.Message.ToString();
             }
-
             var items = GetFiles();
             return View(items);
         }
 
+        /* --- List of files --- */
+        /** 
+            Summary function which looking for files in choosen directory and return the string List. 
+            Used to display file names on the page. 
+        **/
         private List<string> GetFiles()
         {
             DirectoryInfo folder = new DirectoryInfo(Path.Combine(this.Environment.ContentRootPath, "Images"));
@@ -68,10 +92,14 @@ namespace Optimage.Controllers
             {
                 images.Add(file.Name);
             }
-
             return images;
         }
 
+        /* --- Download files --- */
+        /** 
+            Result is generated for every uploaded file as an action link.
+            Return action for file download.
+        **/
         public FileResult Download(string ImageName)
         {
             var FileVirtualPath = Path.Combine(this.Environment.ContentRootPath, "Images") + "\\" + ImageName;
